+++ NEWS.md
@@ -40,7 +40,8 @@ Version 0.11b0 - dev
 - NEW: `Polyline.append_formatted_points()` with user defined point format support
 - NEW: `size` and `center` properties for bounding box classes
 - NEW: `Insert.ucs()` returns an UCS placed in block reference `insert` location, UCS axis aligned to the block axis.
-- NEW: `Insert.reset_transformation()` reset block reference location, rotation and extrusion vector. 
+- NEW: `Insert.reset_transformation()` reset block reference location, rotation and extrusion vector.
+- NEW: support for associating HATCH boundary paths to geometry entities 
 - BUGFIX: fixed base point calculation of aligned dimensions
 - BUGFIX: fixed length extension line support for linear dimensions
 - BUGFIX: `UCS.to_ocs_angle_deg()` and `UCS.to_ocs_angle_rad()`